+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
-_No unreleased changes yet._
+- refactor: renaming main folder --> `gha-add-update-label-weekly` to accommodate fut. ghas
+- refactor: remove overly specific ref to template in `resolve-config.js`
+
 
 ## v0.6.0
 2025-12-03