+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+_No unreleased changes yet._
+
+## v0.6.2
+2026-01-10
 - feature: Added simplified, consolidated logging for each issue via `issueLogger()`
 - bug: Fixed major versioning